Hibernate Working Environment
This commit has the Hibernate working environment. The hibernate is setup based on the Hibernate ORM 4.2. This doesn't use the Hibernate Search Zip file downloaded, but this works fine. This is a very simple set up for Hibernate. This is tested for addding objects into the database. The database name should be "SmartHomeSystem".